feat : 생성된 코드를 사용자가 읽기 쉬운 형태로 제공한다

+++ strategy-service/src/main/java/com/pda/strategy_service/service/StrategyCodeGenerator.java
@@ -14,7 +14,7 @@ import java.util.*;
 public class StrategyCodeGenerator {
 
   private Set<String> timeframes;
-  private int memberId;
+  private Long memberId;
   private String symbol;
 
   /**
@@ -56,7 +56,7 @@ public class StrategyCodeGenerator {
   /**
    * 전략 JSON → Python 코드 생성
    */
-  public String generateCode(Map<String, Object> strategyJson, int memberId, String symbol) {
+  public String generateCode(Map<String, Object> strategyJson, Long memberId, String symbol) {
     this.timeframes = analyzeTimeframes(strategyJson);
     this.memberId = memberId;
     this.symbol = symbol;
